arm: clarify comment about exception stack use
Each mode has its own stack, but we only use it as a base pointer
to a private memory region. That region, which has a size of
sizeof(struct pt_regs), is used to store some registers during the
switch to svc mode - which will use the normal svc stack. The switch
to svc mode is done because we handle all exceptions in svc.

Signed-off-by: Andrew Jones <drjones@redhat.com>